--- a/dejafu/Test/DejaFu/Internal.hs
+++ b/dejafu/Test/DejaFu/Internal.hs
@@ -135,7 +135,7 @@ tvarsOf act = tvarsRead act `S.union` tvarsWritten act
 -- didn't @retry@).
 tvarsWritten :: ThreadAction -> Set TVarId
 tvarsWritten act = S.fromList $ case act of
-  STM trc _ -> concatMap tvarsOf' trc
+  STM trc _ _ -> concatMap tvarsOf' trc
   BlockedSTM trc -> concatMap tvarsOf' trc
   _ -> []
 
@@ -149,7 +149,7 @@ tvarsWritten act = S.fromList $ case act of
 -- | Get the @TVar@s a transaction read from.
 tvarsRead :: ThreadAction -> Set TVarId
 tvarsRead act = S.fromList $ case act of
-  STM trc _ -> concatMap tvarsOf' trc
+  STM trc _ _ -> concatMap tvarsOf' trc
   BlockedSTM trc -> concatMap tvarsOf' trc
   _ -> []
 
@@ -189,12 +189,12 @@ rewind (ModIORefCas c) = WillModIORefCas c
 rewind (WriteIORef c) = WillWriteIORef c
 rewind (CasIORef c _) = WillCasIORef c
 rewind (CommitIORef t c) = WillCommitIORef t c
-rewind (STM _ _) = WillSTM
+rewind (STM _ _ _) = WillSTM
 rewind (BlockedSTM _) = WillSTM
 rewind Catching = WillCatching
 rewind PopCatching = WillPopCatching
-rewind (Throw _) = WillThrow
-rewind (ThrowTo t _) = WillThrowTo t
+rewind (Throw _ _) = WillThrow
+rewind (ThrowTo t _ _) = WillThrowTo t
 rewind (BlockedThrowTo t) = WillThrowTo t
 rewind (SetMasking b m) = WillSetMasking b m
 rewind (ResetMasking b m) = WillResetMasking b m
@@ -295,8 +295,8 @@ tidsOf (TryPutMVar _ _ tids) = S.fromList tids
 tidsOf (TakeMVar _ tids) = S.fromList tids
 tidsOf (TryTakeMVar _ _ tids) = S.fromList tids
 tidsOf (CommitIORef tid _) = S.singleton tid
-tidsOf (STM _ tids) = S.fromList tids
-tidsOf (ThrowTo tid _) = S.singleton tid
+tidsOf (STM _ tids _) = S.fromList tids
+tidsOf (ThrowTo tid _ _) = S.singleton tid
 tidsOf (BlockedThrowTo tid) = S.singleton tid
 tidsOf _ = S.empty
 
@@ -374,8 +374,11 @@ updateMaskState tid (Fork tid2) = \masks -> case M.lookup tid masks of
   Nothing -> masks
 updateMaskState tid (SetMasking   _ ms) = M.insert tid ms
 updateMaskState tid (ResetMasking _ ms) = M.insert tid ms
-updateMaskState tid (Throw True) = M.delete tid
-updateMaskState _ (ThrowTo tid True) = M.delete tid
+updateMaskState tid (Throw _ True) = M.delete tid
+updateMaskState tid (Throw (Just ms) _) = M.insert tid ms
+updateMaskState _ (ThrowTo tid _ True) = M.delete tid
+updateMaskState _ (ThrowTo tid (Just ms) _) = M.insert tid ms
+updateMaskState tid (STM _ _ (Just ms)) = M.insert tid ms
 updateMaskState tid Stop = M.delete tid
 updateMaskState _ _ = id