--- a/dejafu/Test/DejaFu/SCT/Internal.hs
+++ b/dejafu/Test/DejaFu/SCT/Internal.hs
@@ -397,8 +397,8 @@ renumber memtype tid0 crid0 = snd . mapAccumL go (I.empty, tid0, I.empty, crid0)
     (s, WriteIORef (renumbered cridmap old))
   updateAction s@(_, _, cridmap, _) (CasIORef old b) =
     (s, CasIORef (renumbered cridmap old) b)
-  updateAction s@(tidmap, _, _, _) (STM tas olds ms) =
-    (s, STM tas (map (renumbered tidmap) olds) ms)
+  updateAction s@(tidmap, _, _, _) (STM tas olds) =
+    (s, STM tas (map (renumbered tidmap) olds))
   updateAction s@(tidmap, _, _, _) (ThrowTo old ms b) =
     (s, ThrowTo (renumbered tidmap old) ms b)
   updateAction s@(tidmap, _, _, _) (BlockedThrowTo old) =