… .hlint.yaml
… dejafu/CHANGELOG.rst
… dejafu/Test/DejaFu/SCT/Internal/DPOR.hs
… doc/getting_started.rst
… hunit-dejafu/CHANGELOG.rst
… tasty-dejafu/CHANGELOG.rst