## Summary

Previously, dejafu did not atomically restore the masking state of a thread when jumping to an exception handler.  This opens a possibility of a race with another thread calling `throwTo`, and this is a race which GHC does not do!

This PR makes the restoration of the masking state atomic, changes `Throw` and `ThrowTo` to report the new masking state (if it changed), and adds a new `ThrownSTM` action.

**Related issues:** fixes #324

## Checklist

- [x] Add new tests to dejafu-tests

## Benchmark results (for performance issues)

**Before:**

```
$ stack exec -- dejafu-tests +RTS -s

 135,743,041,632 bytes allocated in the heap
  22,764,435,640 bytes copied during GC
      13,045,840 bytes maximum residency (4468 sample(s))
          96,712 bytes maximum slop
              12 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     126630 colls,     0 par   14.163s  14.146s     0.0001s    0.0008s
  Gen  1      4468 colls,     0 par    7.858s   7.865s     0.0018s    0.0162s

  TASKS: 5 (1 bound, 4 peak workers (4 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time   70.529s  ( 67.576s elapsed)
  GC      time   22.021s  ( 22.011s elapsed)
  EXIT    time    0.000s  (  0.002s elapsed)
  Total   time   92.550s  ( 89.590s elapsed)

  Alloc rate    1,924,643,085 bytes per MUT second

  Productivity  76.2% of total user, 75.4% of total elapsed
```

**After:** (with the two new tests commented out)

```
$ stack exec -- dejafu-tests +RTS -s

 138,271,002,848 bytes allocated in the heap
  23,508,977,296 bytes copied during GC
      13,335,744 bytes maximum residency (4686 sample(s))
         128,928 bytes maximum slop
              12 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     128854 colls,     0 par   14.606s  14.577s     0.0001s    0.0008s
  Gen  1      4686 colls,     0 par    7.999s   7.993s     0.0017s    0.0146s

  TASKS: 5 (1 bound, 4 peak workers (4 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time   71.915s  ( 68.950s elapsed)
  GC      time   22.605s  ( 22.571s elapsed)
  EXIT    time    0.000s  (  0.010s elapsed)
  Total   time   94.521s  ( 91.530s elapsed)

  Alloc rate    1,922,689,157 bytes per MUT second

  Productivity  76.1% of total user, 75.3% of total elapsed
```

I was worried the extra allocation may be a problem, but it's fine.