@mitchellwrosen 2020-06-27
Looks great, with this patch, my (one) test does behave as expected :)

```haskell
Throw (Maybe MaskingState) Bool
￼   -- ^ Throw an exception, and give the resultant masking state after
￼   -- jumping to the exception handler (if it changed).  If the 'Bool'
￼   -- is @True@, then this killed the thread.
￼```

What does it mean to get `Just`+`True` here, or is that not possible?