@barrucadu 2020-06-27
Yes, `Just` = `True` is impossible.  Do you think I should change the type?  Or would clarifying the haddock be enough?