@mitchellwrosen 2020-06-28
I kind of like making only the valid states representable, but it's totally up to you.

Perhaps the field could be reworked to "the resultant masking state after jumping to the exception handler (if any)" - then `Nothing` would mean the thread died, and there'd be no need for the separate bool.